requirements.txt
streamlit==1.32.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.2.1
pillow==10.2.0
scikit-learn==1.7.2
scipy==1.15.3
matplotlib==3.8.4